MariaDB: Improve README.md
Reorders some of the content to make the production vs. quickstart
more obvious.

Recommend the use of <yourname>/mariadb for the creation of the
data volume so as not to force the user to pull down the 'fedora'
image if they don't already have it.

Add --protocol=tcp to the mysqladmin and mysql commands, as
otherwise they will attempt to talk to the local socket which is
not connected to the container.