mariadb: simplify: 'root'@'localhost' password serves no purpose
mysqld is the only process on localhost.

Unprivileged remote mysql users should not be able to spoof as localhost.
If they can, something has gone very wrong, and it's quite possible they
could bypass the mysql root password anyway.

It might still be useful to have a separate root user, and it doesn't look
like it hurts to leave it around.

Closes issue #190.