mariadb: maybe we shouldn't run mysqld as root
Upgrading to this image is likely to break due to file permissions.
As discussed in github issues, Fedora-Dockerfiles is expected to be used
for learning purposes, not directly in production.  There's a similar
breaking change I really want to get out there (set a fixed UID), and I'm
not eager to add a `chown` command because it would defeat the point of
that.

We don't need to put USER in the Dockerfile.  Actually for learning
purposes it's much friendlier if users can run bash in the container
without remembering to pass `-u root` to docker.  And also be able to run
our scripts as root without causing permissions problems.  This is the
approach used by "official" images on the docker hub.
http://stackoverflow.com/questions/36137953/why-do-docker-official-images-not-use-user-as-required-by-best-practices/36137954#36137954

This oversight was probably because a) it's common to have config files
that tell mysqld to drop privileges b) Fedora doesn't, and relies on
systemd starting it as the user mysql instead.  People should really read
the .service files when they port to Docker :-P.