mariadb: use docker for logging
This requires

 1. Add custom mariadb-server.cnf
 2. Set user= instead of passing -u each time
 3. Drop mysqld_safe.  It doesn't support logging to stdout/stderr (really).
 4. Finally, drop log-file= to enable logging to stdout/stderr

Frankly it's arguable. The first 3 points aren't bad. And it means we show
errors directly, instead of hiding them somewhere. The disadvantage is that
setting up mysqld is very noisy; mysqld is started *four times* when you
first run this container.

I think the main argument for pushing this is that we're making a nice
example container for popular software, showing how you can follow the
docker ideals / "best practice" document etc. (Albeit the nice-ness is
limited by how nice the software is to start with). It'd be a shame to stop
half-way without a specific reason.