mariadb: set fixed UID for mysql user
This is my second breaking change.

This aims to promote the idea of fixed UIDs.  The alternative to fixing
UIDs is to add a script which runs chown, to handle when the base image
gains a new user.  There's nothing particularly wrong with that for mysqld,
but it becomes sub-optimal as volumes grow to have large numbers of files.
Paying attention to UIDs is the "correct" answer to that.  I've not seen
anyone suggest a drawback to it, except that it means choosing a uid.

"Add our user and group first to make sure their IDs get assigned
consistently" returns 900+ results on google.  People think they're doing
the right thing.  However we don't have documentation to say Fedora base
images are guaranteed to allocate exactly N UIDs.

Some images use uid 999, because on debian (`useradd`) this is the last
"system" uid to be allocated.  On Fedora (`adduser`), uid 999 is the
first to be allocated.  So uid 1000 is a less confusing choice.

If the uid is already allocated, we should see the build fail with an error
that mentions uids.  As opposed to failing at runtime with a blanket
"Permission denied" :-).