A reworked version of #194
These changes are based on the code that from https://github.com/sclorg/mariadb-container.
Even if the image is meant to work in OpenShift, it doesn't mean that it shouldn't
work outside of OpenShift -- the opposite is true, the image should be usable on
Atomic or normal Fedora Server/Workstation/Cloud without any issues as well.